UPSTREAM: drivers/generic/gpio_keys: Add trigger for wakeup event action
Currently without any trigger the wakeup event is generated on both the
rising and falling edges of the GPIO input. Add support to specify the
trigger explicitly so that the configuration can be passed to the
kernel.

BRANCH=octopus
BUG=b:117953118
TEST=Ensure that the system boots to ChromeOS. Ensure that the stylus tools
open on pen eject. Ensure that the system wakes on Pen Eject. Ensure that
the system enters S0ix and S3 states after the pen is ejected. Ensure that
the system enters S0ix and S3 states when the pen remains inserted in its
holder. Ensured that the system does not wake when the pen is inserted.
Ensure that the suspend_stress_test runs successfully for 25 iterations
with the pen placed in its holder and ejected from its holder.

Change-Id: I58bcaf3d98c1aabbce685daba5d33660fbdc8a67
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c4427393c5c7e75c60b2e91113adb7ca0ec94f8f
Original-Change-Id: Ifb08ba01106031aa2655c1ae2faab284926f1ceb
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/30451
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1434502
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>