UPSTREAM: mb/google/slippy: Add a VBT for all variants
The vbt was extracted from the option rom found on stock images.
The vbt.bin is the same across all variants.

The VBT has a modified BDB block 43, the 'Backlight info block' such
that the inverter type for the panel in use is set to
2 (BDB_BACKLIGHT_TYPE_PWM) instead of 0 (BDB_BACKLIGHT_TYPE_NONE).
This only seems to matter on Windows, as without it changing the
backlight duty cycle does not work.

BUG=none
BRANCH=none
TEST=none

Change-Id: I27a7b6b703331c0be9e74e72e08481e666c4ed00
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d2f678d3bd6ca4c05fa5c652d6cdf4623543e576
Original-Change-Id: I82c72c561e1058e0b77d80baf330b64f7c6b08e3
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/30487
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Reviewed-by: Tristan Corrick <tristan@corrick.kiwi>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1397730
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>