UPSTREAM: sb/intel/common: Show "Add EC firmware" only for boards that need it
Most boards currently do not use EC firmware from SPI flash in the
IFD, this hides this option by default and shows it only for boards
that need it.

A new config variable MAINBOARD_USES_IFD_EC_REGION is introduced to
enable this option for boards that need it.

The following list of boards requiring this was provided by
Lijian Zhao:
1. intel/cannonlake_rvp
2. intel/coffeelake_rvp
3. intel/icelake_rvp
4. google/sarien
5. google/hatch

BUG=none
BRANCH=none
TEST=none

Change-Id: I5dc0dd2dd4da55cbe1f123104f40e3670399012a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 82a4e27341bf292925d25779985a38f70a9858aa
Original-Change-Id: I52ab977319d99a23a5e982cc01479fe801e172a7
Original-Signed-off-by: Jan Tatje <jan@jnt.io>
Original-Reviewed-on: https://review.coreboot.org/c/30697
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1405000
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>