UPSTREAM: soc/intel/cannonlake: Fix chipset_power_state structure
This structure is declared as a static CAR_GLOBAL in the common
PMC library code and in the SOC specific code.  Remove the SOC
specific version and instead get the chipset_power_state pointer
from the PMC library.

This fixes events that were recorded in chipset_power_state at
boot but were reading as all zero when it was time to parse the
structure when logging events to flash.

BUG=none
BRANCH=none
TEST=none

Change-Id: I66b23e06b7f3d91dba84cc25a409936e5eadfe92
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3da1b0d439f249a3e4a056ba24890688adb88d4d
Original-Change-Id: I67a4f724c0707d98766ad28abd8d0b66a5615745
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30717
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1402673
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>