UPSTREAM: mb/google/sarien: Enable recovery mode GPIO
Enable the active-low recovery mode GPIO now that new boards are
available which have an external pull-up instead of a pull-down so
it can be asserted properly by servo.

This was tested on a Sarien system by holding the recovery button
on the servo board and tapping the cold reset button and ensuring
that it enters recovery mode.

BUG=none
BRANCH=none
TEST=none

Change-Id: I45d59c16b3f6250719723e628a922f53d08d9b27
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: cae7944fc310be9ef6b0c105d5a6a3e9b113e507
Original-Change-Id: I3216580bc94de71b05bf9382f15d0c4d428cb9fa
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30721
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1402677
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>