UPSTREAM: soc/intel/cannonlake: Add FSP UPD for minimum assertion width
Expose the FSP tunables for the chipset minimum assertion width
settings which can be configured per-board.

The defaults appear to be different from what is listed in the FSP
header documentation so I tried to list what the actual default is
based on the source rather than what is stated the header comments.

BUG=none
BRANCH=none
TEST=none

Change-Id: I00703cbb1f6c56787896191eb45ee6db37c08adf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 55012d149ae3f917da419a2987081739222fc357
Original-Change-Id: Ie0606c2984727adf13c9fb8395586287162e49ca
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30719
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lijian Zhao <lijian.zhao@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/1402675
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>