UPSTREAM: soc/intel/baytrail: Use non-evict CAR setup
The CAR setup is almost identical to the cpu/intel/non-evict
CAR setup, with the only difference that L2 cache needs to be
separately enabled. Currently this assumes that it is possible
to use a static Kconfig option to cover all CPU's requiring this.

BUG=none
BRANCH=none
TEST=none

Change-Id: Iad5fe9e8fb437c970d470bf9b730f4dd5cc5a09c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b1c57d1bebb6dd516afa2e85a3f8082a6c77f8ec
Original-Change-Id: Iae9b584bc0d32a56be2e6e2b2e893897eb448aa5
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/30814
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1463057
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>