UPSTREAM: soc/intel/fsp_broadwell_de: Fix TSEG size computation
The address bits 19:0 of TSEG_LIMIT read as zero, but are ignored on
comparison. The result is that the limit is effectively FFFFFh.

Add one MiB to the register value to make TSEG 8MiB instead of 7MiB.
Fixes a crash related to SMRR not matching the TSEG region.

BUG=none
BRANCH=none
TEST=none

Change-Id: I54b5813527d4f3f68eb134e80d1cd263e4f57175
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7ace555cc112c6942814eb28496c8d9cd119da78
Original-Change-Id: I1a625f7bb53a3e90d3cbc0ce16021892861367d8
Original-Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/30932
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/1418800
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>