UPSTREAM: Documentation: Add Project Ideas document
We already had such a page on the wiki, but it's outdated and the wiki
is supposed to go the way of the dodo anyway.

This is a fresh start to make sure that all ideas we're coming up with
are still current and that there are mentors willing to support them.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id78db4a51f3b478d1755f20aba369a41ac5fdfe3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8fe2d5403fecc9d780870ac8288dd10db93112c9
Original-Change-Id: Idd68f845930bd37a2293969b9a153cf584d6d15f
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/30972
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1455913
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>