UPSTREAM: src/soc/intel/cannonlake: Add _DSM methods for LPIT table
This patch adds the _DSM method 5 and 6 for entering and exiting S0ix.
The _DSM method gets injected into DSDT table and called from kernel.

LPIT table is hardcoded in this patch but the proper way to implement
is to use inject_dsdt to make the _DSM methods available for soc's to
implement.

Calling the LPIT table from mainboard here so that with the current
implementation the platforms which do not have lpit support throw
compilation error.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia9d5ff5def55c93c82148417d21c24e2dc91d852
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5620b105461cc18cf1439f02013153237f372b4b
Original-Signed-off-by: Lijian Zhao <lijian.zhao@intel.com>
Original-Change-Id: Ia908969decf7cf12f505becb4f4a4a9caa7ed6db
Original-Reviewed-on: https://review.coreboot.org/c/31101
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Shaunak Saha <shaunak.saha@intel.corp-partner.google.com>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1480524
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>