UPSTREAM: mb/google/octopus: Add USB ACPI configuration for CNVi BT module
This change enables exporting the reset GPIO for CNVi Bluetooth module to
the kernel for use in an rf-kill operation.

BUG=b:123296264
BRANCH=octopus
TEST=Boots to ChromeOS. Checked the SSDT table to ensure that the reset
gpio is exported under the device \_SB_.PCI0.XHCI.RHUB.HS09. Ensured
that the kernel btusb driver is able to find the exported GPIO in the
devices with CNVi BT module.

Change-Id: I7eed8275467a997a39fa728b5663abec8bd32922
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8a64b6fbdbbf32dafe84ff8d8c74a3a79f1ae169
Original-Change-Id: I10f28bfe705da5104d709ae2ed91a8ae003fa639
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/31183
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1454390
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>