UPSTREAM: mb/google/poppy/variants/atlas: move WiFi wake to GPP_B7
The latest rev. of the atlas board moves the WiFi wake source from
WAKE# to GPP_B7.  The original GPP_A0 in the device tree is just
wrong.  This also reconfigures DW1 to the GPP_B group so we can use
GPP_B7 as a wake source.

GPP_B7 is still configured as a no-connect in gpio.c, so this doesn't
actually enable WiFi wake.  We'll follow up with another patch to set
up GPP_B7 properly on boards that support it.

BUG=b:122327852
BRANCH=none
TEST=atlas still boots

Change-Id: I37e7f12c94c363edc9fd00c4e67cf6cacd142bd1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id:
Original-Change-Id: If34f98dc0b1cd829ba731608a4547c6284af58af
Original-Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Original-Commit-Id: ab770083952dbe39987650a9570967b1436ce785
Original-Original-Change-Id: I1816500dd0ab6186fd51aa6945faf73d00c152fe
Original-Original-Signed-off-by: Caveh Jalali <caveh@chromium.org>
Original-Original-Reviewed-on: https://review.coreboot.org/c/31211
Original-Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1485891
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>