UPSTREAM: drivers/gpio_keys: Remove redundant is_wakeup_source flag
"is_wakeup_source" flag is used to indicate if the concerned device can
trigger a wakeup. This flag is redundant with the "wake" GPE event
definition. So remove the redundant flag and use the "wake" GPE event to
mark the wakeup source.

BUG=None
BRANCH=None
TEST=Boot to ChromeOS. Ensure that the device is marked as wakeup-source
in SSDT if wake GPE is configured. Ensure that the system can suspend
and the device acts as a wakeup source

Change-Id: I31ab8ccaed68fc8251e8b0d370a867d2118f13f8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d5c458f98ad71f66815f637c5cb6305adc133b27
Original-Change-Id: I99237323639df1cb72e3a81bcfed869900a2eefa
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Reviewed-on: https://review.coreboot.org/c/31258
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1460848
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>